[mod] confirm delete sub ques

+++ vue-ui/src/components/adminTemplate/AdminTemplateSubQuestion.vue
@@ -109,10 +109,12 @@ export default {
       })
     },
     deleteSubQues(id) {
-      this.$http.delete(`/api/admin/questionnaire/${this.templateId}/question/${this.questionId}/sub/${id}`).then(res => {
-        if (res.status == 200) {
-          this.subQuestionList = res.data.subQuestionList
-        }
+      this.$confirm('Confirm delete?',{type: 'warning'}).then(() => {
+        this.$http.delete(`/api/admin/questionnaire/${this.templateId}/question/${this.questionId}/sub/${id}`).then(res => {
+          if (res.status == 200) {
+            this.subQuestionList = res.data.subQuestionList
+          }
+        })
       })
     },
     writeSubQues(id,vid) {
